pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wpf-agent"
version = "0.1.0"
description = "WPF UI Debug Automation Agent - Claude Code MCP integration"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pywinauto>=0.6.8",
    "mcp>=1.0.0",
    "click>=8.1",
    "pydantic>=2.0",
    "Pillow>=10.0",
    "psutil>=5.9",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
build = [
    "pyinstaller>=6.0",
]

[project.urls]
Homepage = "https://github.com/shiro-mac/wpf-agent"

[project.scripts]
wpf-agent = "wpf_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wpf_agent"]

[tool.hatch.build.targets.wheel.force-include]
".claude/skills" = "wpf_agent/_skills"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
